src/asttroshield/udl_integration/requirements.txt
requests>=2.28.0
confluent-kafka>=2.0.0
python-dateutil>=2.8.2
urllib3>=1.26.0
python-dotenv>=1.0.0
pytest>=7.0.0
mock>=5.0.0
kafka-python>=2.0.0
threading2>=0.3.1 